feat: Add mongo_dart and http dependencies
The code changes include adding the mongo_dart and http dependencies to the pubspec.yaml file. These dependencies will be used to interact with a MongoDB database and make HTTP requests, respectively. This addition enhances the functionality of the application by enabling data storage and retrieval as well as communication with external APIs.